requirements.txt
streamlit==1.48.1
pandas==2.3.1
matplotlib==3.10.5
joblib==1.5.1
numpy==2.3.2
XGBoost==3.0.4
seaborn==0.13.2
statsmodels==0.14.5
scikit-learn==1.7.1
pyarrow.parquet==21.0.0
mlflow ==3.3.1
